+++ test/mysql-gateway-java-client.mocha.js
@@ -3,19 +3,24 @@ const { spawn, exec } = require('child_process');
 const { readdirSync } = require('fs');
 const { sane } = require('./utils.js');
 
-const TEST_PORT = 3307;
+const TEST_PORT = 13307;
 
 var child;
 
 var jars = readdirSync('test/jdbc/jar');
+if (!jars.length) throw new Error("must have at least one jar in 'test/jdbc/jar'");
 
-describe('mysql-gateway-mysql-client', () => {
+// jars = ['mysql-connector-java-6.0.2.jar'];
+
+describe('mysql-gateway-mysql-client', function() {
+  this.timeout(10000);
+
   before((done) => {
     exec('javac test/jdbc/DruidQuery.java', (error, stdout, stderr) => {
       expect(error).to.equal(null);
       expect(stderr).to.equal('');
 
-      child = spawn('bin/plyql', `-h 192.168.99.100 -i P2Y --experimental-mysql-gateway ${TEST_PORT}`.split(' '));
+      child = spawn('bin/plyql', `-h 192.168.99.100 -i P2Y --druid-time-attribute time --force-boolean isNew --experimental-mysql-gateway ${TEST_PORT}`.split(' '));
 
       child.stderr.on('data', (data) => {
         throw new Error(data.toString());
@@ -23,6 +28,7 @@ describe('mysql-gateway-mysql-client', () => {
 
       child.stdout.on('data', (data) => {
         data = data.toString();
+        //console.log('GATEWAY:', data);
         if (data.indexOf(`port: ${TEST_PORT}`) !== -1) {
           done();
         }
@@ -33,30 +39,37 @@ describe('mysql-gateway-mysql-client', () => {
 
   for (var jar of jars) {
     it(`does basic query on ${jar}`, (testComplete) => {
-      exec(`java -cp test/jdbc/jar/${jar}:test/jdbc/ DruidQuery`, (error, stdout, stderr) => {
+      exec(`java -cp test/jdbc/jar/${jar}:test/jdbc/ DruidQuery "jdbc:mysql://127.0.0.1:${TEST_PORT}/plyql1"`, (error, stdout, stderr) => {
         expect(error).to.equal(null);
+        //console.log('stderr', stderr);
         expect(stdout).to.contain(sane`
-          page[Jeremy Corbyn] count[314]
-          page[User:Cyde/List of candidates for speedy deletion/Subpage] count[255]
-          page[Wikipedia:Administrators' noticeboard/Incidents] count[228]
-          page[Wikipedia:Vandalismusmeldung] count[186]
-          page[Total Drama Presents: The Ridonculous Race] count[160]
-          page[Wikipedia:Administrator intervention against vandalism] count[145]
-          page[Flavia Pennetta] count[141]
-          page[Wikipedia:Requests for page protection] count[132]
-          page[Wikipédia:Le Bistro/12 septembre 2015] count[130]
-          page[List of shipwrecks in August 1944] count[129]
-          page[Wikipedia:Auskunft] count[127]
-          page[Wikipedia:In the news/Candidates] count[126]
-          page[Wikipedia:Prośby o przejrzenie edycji] count[126]
-          page[Wikipedia:Files for deletion/2015 September 12] count[122]
-          page[Wikipedia:Adminkandidaturen/Nicola III] count[115]
-      `);
+          Time[2015-09-11 17:00:00.0] Channel[en] Count[104870] Added[231369.562500]
+          Time[2015-09-11 17:00:00.0] Channel[vi] Count[98862] Added[29220.669922]
+          Time[2015-09-11 17:00:00.0] Channel[de] Count[23833] Added[40796.171875]
+          Time[2015-09-11 17:00:00.0] Channel[fr] Count[19064] Added[41101.738281]
+          Time[2015-09-11 17:00:00.0] Channel[ru] Count[12841] Added[34958.781250]
+        `);
         //expect(stderr).to.equal('');
         testComplete();
       });
     });
   }
+
+  it(`does same query on real MySQL ${jar}`, (testComplete) => {
+    exec(`java -cp test/jdbc/jar/${jar}:test/jdbc/ DruidQuery "jdbc:mysql://192.168.99.100:3306/datazoo?user=root"`, (error, stdout, stderr) => {
+      expect(error).to.equal(null);
+      //console.log('stderr', stderr);
+      expect(stdout).to.contain(sane`
+        Time[2015-09-11 17:00:00.0] Channel[en] Count[104870] Added[231369.562500]
+        Time[2015-09-11 17:00:00.0] Channel[vi] Count[98862] Added[29220.669922]
+        Time[2015-09-11 17:00:00.0] Channel[de] Count[23833] Added[40796.171875]
+        Time[2015-09-11 17:00:00.0] Channel[fr] Count[19064] Added[41101.738281]
+        Time[2015-09-11 17:00:00.0] Channel[ru] Count[12841] Added[34958.781250]
+      `);
+      //expect(stderr).to.equal('');
+      testComplete();
+    });
+  });
 
   after(() => {
     child.kill('SIGHUP');